Move compare_* from test_helpers to query_comparison.py
