+++ graphql_compiler/tests/test_compiler.py
@@ -14,10 +14,8 @@ from ..compiler import (
 )
 from ..compiler.ir_lowering_sql.metadata import SqlMetadata
 from .test_data_tools.data_tool import get_animal_schema_sql_metadata
-from .test_helpers import (
-    SKIP_TEST, compare_cypher, compare_gremlin, compare_input_metadata, compare_match, compare_sql,
-    get_schema
-)
+from .test_helpers import (SKIP_TEST,  compare_input_metadata, get_schema)
+from ..query_comparison import compare_cypher, compare_gremlin, compare_match, compare_sql
 
 
 def check_test_data(
