+++ graphql_compiler/tests/test_emit_output.py
@@ -19,7 +19,8 @@ from ..compiler.ir_lowering_match.utils import CompoundMatchQuery
 from ..compiler.match_query import convert_to_match_query
 from ..compiler.metadata import LocationInfo, QueryMetadataTable
 from ..schema import GraphQLDateTime
-from .test_helpers import compare_cypher, compare_gremlin, compare_match, get_schema
+from .test_helpers import get_schema
+from ..query_comparison import compare_cypher, compare_gremlin, compare_match
 
 
 class EmitMatchTests(unittest.TestCase):
