+++ graphql_compiler/tests/test_end_to_end.py
@@ -14,7 +14,8 @@ from ..exceptions import GraphQLInvalidArgumentError
 from ..query_formatting import insert_arguments_into_query
 from ..query_formatting.common import validate_argument_type
 from ..schema import GraphQLDate, GraphQLDateTime, GraphQLDecimal
-from .test_helpers import compare_gremlin, compare_match, get_schema
+from .test_helpers import get_schema
+from ..query_comparison import compare_gremlin, compare_match
 
 
 EXAMPLE_GRAPHQL_QUERY = '''{
