+++ graphql_compiler/tests/test_ir_lowering.py
@@ -21,7 +21,8 @@ from ..compiler.ir_lowering_match.utils import BetweenClause, CompoundMatchQuery
 from ..compiler.match_query import MatchQuery, convert_to_match_query
 from ..compiler.metadata import LocationInfo, QueryMetadataTable
 from ..schema import GraphQLDate, GraphQLDateTime
-from .test_helpers import compare_ir_blocks, get_schema
+from .test_helpers import get_schema
+from ..query_comparison import compare_ir_blocks
 
 
 def check_test_data(test_case, expected_object, received_object):
